+++ os/src/driver/led/led.h
@@ -10,11 +10,11 @@
 
 #include "../driver.h"
 
-extern int LEDInit	(short id);
-extern int LEDOpen	(short id);
-extern int LEDClose	(short id);
-extern int LEDWrite	(short id, char* buf, int len);
-extern int LEDRead	(short id, char* buf, int len);
-extern int LEDIoctl	(short id, int cmd, int mode, char* buf, int len);
+extern int LEDInit	(uint16_t id);
+extern int LEDOpen	(uint16_t id);
+extern int LEDClose	(uint16_t id);
+extern int LEDWrite	(uint16_t id, char* buf, uint16_t len);
+extern int LEDRead	(uint16_t id, char* buf, uint16_t len);
+extern int LEDIoctl	(uint16_t id, uint16_t cmd, uint8_t mode, char* buf, uint16_t len);
 
 #endif /* DRIVER_LED_LED_H_ */
